Refactored and rewrote FIleSystem interface.
I have totally redone the entire interaction between storage and computation. Files to be uploaded to S3 no longer are written to disk first. The FileSystem interface now uses BufferedReaders and BufferedWriters rather than File-bound streams, which should make it portable enough that it can be extended to databases and other input types. I've also now reduced the size of the JpAws package substantially, so that it is clear to anyone who wishes to add another cloud compute/storage provider that it is not a monumental task to do so. 

Now that I've fully decoupled storage, the next step in improving design is to decouple computation.